fix: move datetime naive-to-aware test to common.py
Both water-level-csv and well-inventory-csv features have the same step
to convert datetime fields from naive to aware and confirm the correct
timezone offset. This commit moves that step implementation to common.py
to avoid code duplication and ensure consistent testing of datetime
handling across both features.
