+++ tests/features/steps/well-inventory-csv.py
@@ -1,9 +1,7 @@
-from datetime import datetime, timedelta
+from datetime import datetime
 
 from behave import given, when, then
 from behave.runner import Context
-
-from services.util import convert_dt_tz_naive_to_tz_aware
 
 
 @given("valid lexicon values exist for:")
@@ -48,23 +46,38 @@ def step_impl(context: Context):
         except ValueError as e:
             raise ValueError(f"Invalid date_time: {row['date_time']}") from e
 
+    if (
+        not hasattr(context, "datetime_fields")
+        or "date_time" not in context.datetime_fields
+    ):
+        context.datetime_fields = ["date_time"]
+    else:
+        context.datetime_fields.append("date_time")
 
+
+# TODO: implement when optional water levels are added to the well inventory csv testing data
 @given(
     'the optional "water_level_date_time" values are valid ISO 8601 timezone-naive datetime strings (e.g. "2025-02-15T10:30:00") when provided'
 )
 def step_impl(context: Context):
     """Verifies that "water_level_date_time" values are valid ISO 8601 timezone-naive datetime strings."""
-    for row in context.rows:
-        if row.get("water_level_date_time", None):
-            try:
-                date_time = datetime.fromisoformat(row["water_level_date_time"])
-                assert (
-                    date_time.tzinfo is None
-                ), f"water_level_date_time should be timezone-naive: {row['water_level_date_time']}"
-            except ValueError as e:
-                raise ValueError(
-                    f"Invalid water_level_date_time: {row['water_level_date_time']}"
-                ) from e
+    pass
+    # for row in context.rows:
+    #     if row.get("water_level_date_time", None):
+    #         try:
+    #             date_time = datetime.fromisoformat(row["water_level_date_time"])
+    #             assert (
+    #                 date_time.tzinfo is None
+    #             ), f"water_level_date_time should be timezone-naive: {row['water_level_date_time']}"
+    #         except ValueError as e:
+    #             raise ValueError(
+    #                 f"Invalid water_level_date_time: {row['water_level_date_time']}"
+    #             ) from e
+
+    # if not hasattr(context, "datetime_fields") or "water_level_date_time" not in context.datetime_fields:
+    #     context.datetime_fields = ["water_level_date_time"]
+    # else:
+    #     context.datetime_fields.append("water_level_date_time")
 
 
 @when("I upload the file to the bulk upload endpoint")
@@ -75,58 +88,58 @@ def step_impl(context: Context):
     )
 
 
-@then(
-    "all datetime objects are assigned the correct Mountain Time timezone offset based on the date value."
-)
-def step_impl(context: Context):
-    """Converts all datetime strings in the CSV rows to timezone-aware datetime objects with Mountain Time offset."""
-    for i, row in enumerate(context.rows):
-        # Convert date_time field
-        date_time_naive = datetime.fromisoformat(row["date_time"])
-        date_time_aware = convert_dt_tz_naive_to_tz_aware(
-            date_time_naive, "America/Denver"
-        )
-        row["date_time"] = date_time_aware.isoformat()
+# @then(
+#     "all datetime objects are assigned the correct Mountain Time timezone offset based on the date value."
+# )
+# def step_impl(context: Context):
+#     """Converts all datetime strings in the CSV rows to timezone-aware datetime objects with Mountain Time offset."""
+#     for i, row in enumerate(context.rows):
+#         # Convert date_time field
+#         date_time_naive = datetime.fromisoformat(row["date_time"])
+#         date_time_aware = convert_dt_tz_naive_to_tz_aware(
+#             date_time_naive, "America/Denver"
+#         )
+#         row["date_time"] = date_time_aware.isoformat()
 
-        # confirm correct time zone and offset
-        if i == 0:
-            # MST, offset -07:00
-            assert date_time_aware.utcoffset() == timedelta(
-                hours=-7
-            ), "date_time offset is not -07:00"
-        else:
-            # MDT, offset -06:00
-            assert date_time_aware.utcoffset() == timedelta(
-                hours=-6
-            ), "date_time offset is not -06:00"
+#         # confirm correct time zone and offset
+#         if i == 0:
+#             # MST, offset -07:00
+#             assert date_time_aware.utcoffset() == timedelta(
+#                 hours=-7
+#             ), "date_time offset is not -07:00"
+#         else:
+#             # MDT, offset -06:00
+#             assert date_time_aware.utcoffset() == timedelta(
+#                 hours=-6
+#             ), "date_time offset is not -06:00"
 
-        # confirm the time was not changed from what was provided
-        assert (
-            date_time_aware.replace(tzinfo=None) == date_time_naive
-        ), "date_time value was changed during timezone assignment"
+#         # confirm the time was not changed from what was provided
+#         assert (
+#             date_time_aware.replace(tzinfo=None) == date_time_naive
+#         ), "date_time value was changed during timezone assignment"
 
-        # Convert water_level_date_time field if it exists
-        if row.get("water_level_date_time", None):
-            wl_date_time_naive = datetime.fromisoformat(row["water_level_date_time"])
-            wl_date_time_aware = convert_dt_tz_naive_to_tz_aware(
-                wl_date_time_naive, "America/Denver"
-            )
-            row["water_level_date_time"] = wl_date_time_aware.isoformat()
+#         # Convert water_level_date_time field if it exists
+#         if row.get("water_level_date_time", None):
+#             wl_date_time_naive = datetime.fromisoformat(row["water_level_date_time"])
+#             wl_date_time_aware = convert_dt_tz_naive_to_tz_aware(
+#                 wl_date_time_naive, "America/Denver"
+#             )
+#             row["water_level_date_time"] = wl_date_time_aware.isoformat()
 
-            if wl_date_time_aware.dst():
-                # MDT, offset -06:00
-                assert wl_date_time_aware.utcoffset() == timedelta(
-                    hours=-6
-                ), "water_level_date_time offset is not -06:00"
-            else:
-                # MST, offset -07:00
-                assert wl_date_time_aware.utcoffset() == timedelta(
-                    hours=-7
-                ), "water_level_date_time offset is not -07:00"
+#             if wl_date_time_aware.dst():
+#                 # MDT, offset -06:00
+#                 assert wl_date_time_aware.utcoffset() == timedelta(
+#                     hours=-6
+#                 ), "water_level_date_time offset is not -06:00"
+#             else:
+#                 # MST, offset -07:00
+#                 assert wl_date_time_aware.utcoffset() == timedelta(
+#                     hours=-7
+#                 ), "water_level_date_time offset is not -07:00"
 
-            assert (
-                wl_date_time_aware.replace(tzinfo=None) == wl_date_time_naive
-            ), "water_level_date_time value was changed during timezone assignment"
+#             assert (
+#                 wl_date_time_aware.replace(tzinfo=None) == wl_date_time_naive
+#             ), "water_level_date_time value was changed during timezone assignment"
 
 
 @then("the response includes a summary containing:")
